ubi-utils: ubinize: add "-c max-leb-count" input parameter
While generating ubi image with ubinize, (almost) all UBI data required for
Fastmap is present. If we generate the Fastmap data as part of the image, the first boot
after flash wount require full scan of the UBI device.
In order to be able to generate the Fastmap data number of device PEBs is required.
A new input parameter was added to ubinize:

 -c, --max-leb-cnt=COUNT  maximum logical erase block count

Change-Id: I066066cc1e53e682a1a69e5498a9af072325f5f7
Signed-off-by: Tanya Brokhman <tlinder@codeaurora.org>