ubi-utils: ubinize: Add fastmap suport to image creation
While generating ubi image with ubinize, all UBI data required for
Fastmap is present. If we generate the Fastmap data as part of the
image, the first boot after flash won't require full scan of the UBI
device.
In order to be able to generate the Fastmap data number of device
PEBs is required. If this input parameter was not provided - Fastmap
will not be generated.

Note: In order for this to work properly change is required in flasher
as well:
If while flashing the image bad blocks were discovered, Fastmap
should be invalidated in the flashed image. Fastmap superblock
(if present) will be in PEB#2 in the image.

Change-Id: I68fa1bb4db4b7cb0b6791634501b108cb11901a3
Signed-off-by: Tanya Brokhman <tlinder@codeaurora.org>